Move implementation of Navigator to NavigatorSession.
